+++ ExpandWorld/data/RoomData.cs
@@ -36,4 +36,6 @@ public class RoomData
   public bool perimeter = false;
   public string size = "";
   public RoomConnectionData[] connections = new RoomConnectionData[0];
+  [DefaultValue(null)]
+  public string[]? objects = null;
 }