+++ publish/CHANGELOG.md
@@ -1,8 +1,14 @@
 - v1.33
   - Adds a new command `ew_dungeons` to print their available rooms.
+  - Adds a blueprint support to the field `objectSpawns` in the `expand_locations.yaml`.
+  - Adds a new field `centerPiece` to the `expand_vegetation.yaml`.
+  - Adds a new field `objects` the `expand_rooms.yaml`.
   - Changes dungeon room names to case insensitive in the `expand_dungeons.yaml` file.
   - Fixes the field `excludeRooms` in the `expand_dungeons.yaml` not working.
   - Fixes Jewelcrafting boss icons not working.
+  - Removes the field `offset` from the `expand_locations.yaml`.
+  - TODO?: swaps working on rooms, first checking room swaps and then location swaps.
+  - TODO?: all in, blueprints as rooms
 
 - v1.32
   - Adds missing location entries automatically to the `expand_locations.yaml` file.